Made an actor helper and started to move some methods in character-sheet into it. For furture expansation - merged the add item and add skill method
